Move asMaxBytesError to error.go (#504)
Extreme nit: since other error functions (including `as*` functions) are
in `error.go`, it seems like it might be worth putting it in there to
avoid an extra file.
